% Generated by roxygen2 (4.0.1): do not edit by hand
\name{applyTransform}
\alias{applyTransform}
\alias{applyTransform.matrix}
\alias{applyTransform.mesh3d}
\title{apply affine transformation to data}
\usage{
applyTransform(x, trafo, inverse)

\method{applyTransform}{matrix}(x, trafo, inverse = FALSE)

\method{applyTransform}{mesh3d}(x, trafo, inverse = FALSE)
}
\arguments{
\item{x}{matrix or mesh3d}

\item{trafo}{4x4 transformation matrix (for mesh3d the matrix will be transformed to a 4x4 matrix)}

\item{inverse}{logical: if TRUE, the inverse of the transformation is applied}
}
\value{
the transformed object
}
\description{
apply affine transformation to data
}
\examples{
data(boneData)
rot <- rotonto(boneLM[,,1],boneLM[,,2])
trafo <- getTrafo4x4(rot)
boneLM2trafo <- applyTransform(boneLM[,,2],trafo)
}

