% Generated by roxygen2 (4.0.1): do not edit by hand
\name{applyTransformation}
\alias{applyTransformation}
\alias{applyTransformation.matrix}
\alias{applyTransformation.mesh3d}
\title{apply affine transformation to data}
\usage{
applyTransformation(x, trafo, inverse)

\method{applyTransformation}{matrix}(x, trafo, inverse = FALSE)

\method{applyTransformation}{mesh3d}(x, trafo, inverse = FALSE)
}
\arguments{
\item{x}{matrix or mesh3d}

\item{trafo}{4x4 transformation matrix}

\item{inverse}{logical: if TRUE, the inverse of the transformation is applied}
}
\value{
the transformed object
}
\description{
apply affine transformation to data
}
\examples{
data(boneData)
rot <- rotonto(boneLM[,,1],boneLM[,,2])
trafo <- getTrafo4x4(rot)
boneLM2trafo <- applyTransformation(boneLM[,,2],trafo)
}

