% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleTest.r
\name{areaSphere}
\alias{areaSphere}
\title{compute the area of an n-dimensional hypersphere}
\usage{
areaSphere(n, r = 1)
}
\arguments{
\item{n}{dimensionality of space the hypersphere is embedded in (e.g.3 for a 3D-sphere)}

\item{r}{radius of the sphere}
}
\value{
returns the area
}
\description{
compute the area of an n-dimensional hypersphere
}
\examples{
areaSphere(2) #gives us the circumference of a circle of radius 1
}
