% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleTest.r
\name{areaSpherePart}
\alias{areaSpherePart}
\title{compute the area of an n-dimensional hypersphere cap}
\usage{
areaSpherePart(n, phi, r = 1)
}
\arguments{
\item{n}{dimensionality of space the hypersphere is embedded in (e.g.3 for a 3D-sphere)}

\item{phi}{angle between vectors defining the cone}

\item{r}{radius of the sphere}
}
\value{
returns the area of the hypersphere cap
}
\description{
compute the area of an n-dimensional hypersphere cap
}
\examples{
areaSpherePart(2,pi/2) # covers half the area of a circle
}
