% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armaGinv.r
\name{armaGinv}
\alias{armaGinv}
\title{calculate Pseudo-inverse of a Matrix using RcppArmadillo}
\usage{
armaGinv(x, tol = NULL)
}
\arguments{
\item{x}{numeric matrix}

\item{tol}{numeric: maximum singular value to be considered}
}
\value{
Pseudo-inverse
}
\description{
a simple wrapper to call Armadillo's pinv function
}
\examples{
mat <- matrix(rnorm(12),3,4)
pinvmat <- armaGinv(mat)
}
