% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2array.r
\name{array2list}
\alias{array2list}
\title{reverts list2array, converting an array to a list of matrices}
\usage{
array2list(x)
}
\arguments{
\item{x}{array}
}
\value{
returns a list containing the matrices
}
\description{
reverts list2array, converting an array to a list of matrices
}
