\docType{data}
\name{boneLM}
\alias{boneLM}
\title{Landmarks and a triangular mesh}
\description{
  Landmarks on the osseous human nose and a triangular mesh representing
  this structure.
}
\usage{boneLM
}
\format{
 An array containing 80 sets of 3D-landmarks placed on the human
    osseous nose.
  
}


  
\keyword{datasets}

