\docType{data}
\name{skull_0144_ch_fe.mesh}
\alias{skull_0144_ch_fe.mesh}

\title{a triangular mesh}
\description{
  triangular mesh representing the osseous nose.
}
\usage{skull_0144_ch_fe.mesh
}
\format{
  The mesh representing the area of the first individual of \code{boneLM} 
}


  
\keyword{datasets}

