\name{c.extract}
\alias{c.extract}

\title{extract information about fixed landmarks, curves and patches from and atlas generated by "landmark"
}
\description{
After exporting the pts file of the atlas from "landmark" and importing it into R via "read.pts" c.extract gets information which 
rows of the landmark datasets belong to curves or patches.
}
\usage{
c.extract(pts.file)
}

\arguments{
  \item{pts.file}{either a character naming the path to a pts.file or the name of an object imported via read.pts.

}
}

\value{returns a list containing the vectors with the indices of matrix rows belonging to the in "landmark" defined curves, patches and fix landmarks.
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{Stefan Schlager

}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{read.pts},\code{\link{read.lmdta}}
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
