\name{cExtract}
\Rdversion{1.1}
\alias{cExtract}

\title{extract information about fixed landmarks, curves and patches from and atlas generated by "landmark"
}
\description{
After exporting the pts file of the atlas from "landmark" and importing it into R via "read.pts" cExtract gets information which 
rows of the landmark datasets belong to curves or patches.
}
\usage{
cExtract(pts.file)
}

\arguments{
  \item{pts.file}{either a character naming the path to a pts.file or the name of an object imported via read.pts.

}
}

\value{returns a list containing the vectors with the indices of matrix rows belonging to the in "landmark" defined curves, patches and fix landmarks.

}

\author{Stefan Schlager

}



\seealso{\code{\link{read.lmdta}} ,\code{\link{read.pts}} }

