% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNA.r
\name{checkNA}
\alias{checkNA}
\title{check for NA values in a matrix (of landmarks)}
\usage{
checkNA(x)
}
\arguments{
\item{x}{matrix containing landmarks}
}
\value{
returns a vector with missin landmarks and a vector of length=0 if none are missing
}
\description{
check for NA values in a matrix (of landmarks)
}
