% Generated by roxygen2 (4.0.1): do not edit by hand
\name{classify}
\alias{classify}
\alias{classify.CVA}
\alias{classify.bgPCA}
\title{classify specimen based on between-group PCA or CVA}
\usage{
classify(x)

\method{classify}{bgPCA}(x)

\method{classify}{CVA}(x)
}
\arguments{
\item{x}{result of groupPCA or CVA}
}
\value{
\item{class}{classification result}
\item{groups}{original grouping variable}
}
\description{
classify specimen based on between-group PCA or CVA
}

