\name{colors}
\alias{bone1}
\alias{bone2}
\alias{bone3}
\alias{skin1}
\alias{skin2}
\alias{skin3}
\alias{skin4}
\docType{data}
\title{
predefined colors for bone and skin}
\description{predefined colors for bone and skin
}
\details{available colors are:

  bone1
  
  bone2
  
  bone3
  
  skin1
  
  skin2
  
  skin3
  
  skin4
  
}
\keyword{datasets}
