% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMissingList.r
\name{createMissingList}
\alias{createMissingList}
\title{create a list with empty entries to be used as missingList in slider3d}
\usage{
createMissingList(x)
}
\arguments{
\item{x}{length of the list to be created}
}
\value{
returns a list of length \code{x} filled with numerics of length zero.
}
\description{
create a list with empty entries to be used as missingList in slider3d
}
\examples{
## Assume in a sample of 10, the 9th individual has (semi-)landmarks 10:50
#   hanging in thin air (e.g. estimated using fixLMtps)
#   while the others are complete.
## create empty list
missingList <- createMissingList(10)
missingList[[9]] <- 10:50
}
\seealso{
\code{\link{fixLMtps},\link{fixLMmirror}, \link{slider3d}}
}
