\name{cSize}
\alias{cSize}
\title{calculate Centroid Size for a landmark configuration}
\usage{
cSize(x)
}
\arguments{
  \item{x}{matrix where each row contains coordinates for
  landmarks}
}
\value{
returns Centroid size
}
\description{
calculate Centroid Size for a landmark configuration
}
\examples{
data(boneData)
cSize(boneLM[,,1])
}
\keyword{~kwd1}
\keyword{~kwd2}

