% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cutMeshPlane}
\alias{cutMeshPlane}
\title{cut a mesh by a hyperplane and remove parts above/below that plane}
\usage{
cutMeshPlane(mesh, v1, v2, v3, keep.upper = TRUE)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{v1}{numeric vector of length=3 specifying a point on the separating plane}

\item{v2}{numeric vector of length=3 specifying a point on the separating plane}

\item{v3}{numeric vector of length=3 specifying a point on the separating plane}

\item{keep.upper}{logical specify whether the points above or below the plane are should be kept}
}
\value{
mesh with part above/below hyperplane removed
}
\description{
cut a mesh by a hyperplane and remove parts above/below that plane
}
\details{
see \code{\link{cutSpace}} for more details.
}

