\name{deprecated}

\alias{c.extract}
\alias{c.size}
\alias{deform.grid}
\alias{mc.CVA}
\alias{mc.find.outliers}
\alias{mc.permudist}
\alias{mc.permuvec}
\alias{mc.procSym}
\alias{mc.slider3d}
\alias{regdist.raw}

\title{
deprecated functions of Morpho}
\description{deprecated functions of Morpho
}

\keyword{internal}
