% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastKmeans.r
\name{fastKmeans}
\alias{fastKmeans}
\title{fast kmeans clustering for 2D or 3D point clouds}
\usage{
fastKmeans(x, k, iter.max = 10, tol = 1e-05,
  threads = parallel::detectCores())
}
\arguments{
\item{x}{matrix containing coordinates or mesh3d}

\item{k}{number of clusters}

\item{iter.max}{maximum number of iterations}

\item{tol}{convergence threshold if the mean distance between centers is below it, the algorithm stops}

\item{threads}{integer number of threads to use}
}
\description{
fast kmeans clustering for 2D or 3D point clouds
}
\examples{
require(Rvcg);require(rgl)
data(humface)
clust <- fastKmeans(humface,k=1000,threads=2)
\dontrun{
require(rgl)
spheres3d(clust$centers)
}
## now look at the vertices closest to the centers
wire3d(humface)
spheres3d(vert2points(humface)[clust$selected,],col=2)
}

