% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetAffineTransform}
\alias{GetAffineTransform}
\title{calculate an affine transformation matrix}
\usage{
GetAffineTransform(x, y, type = c("affine", "rigid", "similarity"))

GetAffineTransform(x, y, type = c("affine", "rigid", "similarity"))
}
\arguments{
\item{x}{fix landmarks}

\item{y}{moving landmarks}

\item{type}{set type of affine transformation: options are "affine", "rigid" and "similarity" (rigid + scale)}

\item{x}{fix landmarks}

\item{y}{moving landmarks}

\item{type}{set type of affine transformation: options are "affine", "rigid" and "similarity" (rigid + scale)}
}
\value{
returns a 4x4 (3x3 in 2D case)  transformation matrix

returns a 4x4 (3x3 in 2D case)  transformation matrix
}
\description{
calculate an affine transformation matrix

calculate an affine transformation matrix
}
\examples{
data(boneData)
trafo <- GetAffineTransform(boneLM[,,1],boneLM[,,2])
transLM <- applyTransform(boneLM[,,2],trafo)
data(boneData)
trafo <- GetAffineTransform(boneLM[,,1],boneLM[,,2])
transLM <- applyTransform(boneLM[,,2],trafo)
}

