% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFaces}
\alias{getFaces}
\title{find indices of faces that contain specified vertices}
\usage{
getFaces(mesh, index)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{index}{vector containing indices of vertices}
}
\value{
vector of face indices
}
\description{
find indices of faces that contain specified vertices
}

