% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls2B.r
\name{getPLSfromScores}
\alias{getPLSfromScores}
\title{compute changes associated with 2-Block PLS-scores}
\usage{
getPLSfromScores(pls, x, y)
}
\arguments{
\item{pls}{output of pls2B}

\item{x}{scores associated with dataset x in original pls2B}

\item{y}{scores associated with dataset y in original pls2B}
}
\value{
returns data in the original space associated with the specified values.
}
\description{
compute changes associated with 2-Block PLS-scores
}
\details{
other than \code{\link{predictPLSfromScores}}, providing Xscores will not compute predictions of y, but the changes in the original data \code{x} that is associated with the specific scores
}
