% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls2B.r
\name{getPLSscores}
\alias{getPLSscores}
\title{compute 2-Block PLS scores for new data}
\usage{
getPLSscores(pls, x, y)
}
\arguments{
\item{pls}{output of pls2B}

\item{x}{matrix or vector representing new dataset(s) -  same kind as in original pls2B}

\item{y}{matrix or vector representing new dataset(s) - same kind as in original pls2B}
}
\value{
returns a vector of pls-scores
}
\description{
compute 2-Block PLS scores for new data from an existing pls2B
}
\note{
either x or y must be missing
}
\seealso{
\code{\link{pls2B}, \link{predictPLSfromScores},\link{predictPLSfromData}}
}
