% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsOnOutline.r
\name{getPointAlongOutline}
\alias{getPointAlongOutline}
\title{Get a point along a line with a given distance from the start of the line}
\usage{
getPointAlongOutline(mat, dist = 15, reverse = FALSE)
}
\arguments{
\item{mat}{matrix with rows containing sequential coordinates}

\item{dist}{numeric: distance from the first point on the line.}

\item{reverse}{logical: if TRUE start from the end of the line}
}
\value{
returns a vector containing the resulting coordinate
}
\description{
Get a point along a line with a given distance from the start of the line
}
