% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotonto.r
\name{getTrafo4x4}
\alias{getTrafo4x4}
\alias{getTrafo4x4.rotonto}
\title{get 4x4 Transformation matrix}
\usage{
getTrafo4x4(x)

\method{getTrafo4x4}{rotonto}(x)
}
\arguments{
\item{x}{object of class "rotonto"}
}
\value{
returns a 4x4 transformation matrix
}
\description{
get 4x4 Transformation matrix
}
\examples{
data(boneData)
rot <- rotonto(boneLM[,,1],boneLM[,,2])
trafo <- getTrafo4x4(rot)
}
