% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotaxis.r
\name{getTrafoRotaxis}
\alias{getTrafoRotaxis}
\title{compute a 4x4 Transformation matrix for rotation around an arbitrary axis}
\usage{
getTrafoRotaxis(pt1, pt2, theta)
}
\arguments{
\item{pt1}{numeric vector of length 3, defining first point on the rotation
axis.}

\item{pt2}{numeric vector of length 3, defining second point on the rotation
axis.}

\item{theta}{angle to rotate in radians. With pt1 being the viewpoint, the
rotation is counterclockwise.}
}
\description{
compute a 4x4 Transformation matrix for rotation around an arbitrary axis
}
\note{
the resulting matrix can be used in \code{\link{applyTransform}}
}
