% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILDS.r
\name{ILDS}
\alias{ILDS}
\title{Compute Pairwise Landmark Distances}
\usage{
ILDS(A)
}
\arguments{
\item{A}{k x m x n array of landmarks}
}
\value{
returns a matrix with each row containing the pairwise distances between landmarks
}
\description{
Compute pairwise landmark distances from landmark array
}
\examples{
require(Morpho)
data(boneData)
proc <- procSym(boneLM)
### compute ILDS from Procrustes aligned landmarks.
edma <- ILDS(proc$rotated)
}
