% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line2plane.r
\name{line2plane}
\alias{line2plane}
\title{get intersection between a line and a plane}
\usage{
line2plane(ptLine, ptDir, planePt, planeNorm)
}
\arguments{
\item{ptLine}{vector of length 3: point on line}

\item{ptDir}{vector of length 3: direction vector of line}

\item{planePt}{vector of length 3: point on plane}

\item{planeNorm}{vector of length 3: plane normal vector}
}
\value{
hit point
}
\description{
get intersection between a line and a plane
}
\note{
in case you only have three points on a plane (named \code{pt1, pt2, pt3} you can get the plane's normal by calling \code{crossProduct(pt1-pt2,pt1-pt3)}.
}
