% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2array.r
\name{list2array}
\alias{list2array}
\title{converts a list of matrices to an array}
\usage{
list2array(x)
}
\arguments{
\item{x}{a list containing matrices of the same dimensionality}
}
\value{
returns an array concatenating all matrices
}
\description{
converts a list of matrices to an array
}
