% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiducials.r
\name{LPS2RAS}
\alias{LPS2RAS}
\title{convert data from LPS to RAS space and back}
\usage{
LPS2RAS(x)
}
\arguments{
\item{x}{mesh of class \code{mesh3d} or a matrix with 3D Landmarks}
}
\value{
returns the rotated data
}
\description{
convert data from LPS to RAS space and back
}
\details{
As e.g. the Slicer versions >= 4.11 are using LPS space, it might be needed to transform data like fiducials and surface models from and back to that space.
}
