\name{mesh2grey}
\alias{mesh2grey}
\title{convert a colored mesh to greyscale.
}
\description{
  convert the colors of a colored mesh to greyscale values
}
\usage{
mesh2grey(mesh)
}
\arguments{
  \item{mesh}{Object of class mesh3d 
}
}

\value{returns a mesh with material$color replaced by greyscale rgb values.

}

\author{
Stefan Schlager}



\seealso{ \code{\link{ply2mesh}},\code{\link{file2mesh}}
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
