% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meshPlaneIntersect}
\alias{meshPlaneIntersect}
\title{get only those faces intersecting a plane}
\usage{
meshPlaneIntersect(mesh, v1, v2, v3)
}
\arguments{
\item{v1}{numeric vector of length=3 specifying a point on the separating plane}

\item{v2}{numeric vector of length=3 specifying a point on the separating plane}

\item{v3}{numeric vector of length=3 specifying a point on the separating plane}
}
\value{
returns a mesh with only those faces intersecting a plane
}
\description{
get only those faces intersecting a plane
}

