% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morpho-package.R
\docType{package}
\name{Morpho-package}
\alias{Morpho-package}
\alias{Morpho}
\title{A toolbox providing methods for data-acquisition, visualisation and
statistical methods related to Geometric Morphometrics and shape analysis}
\description{
A toolbox for Morphometric calculations. Including sliding operations for
Semilandmarks, importing, exporting and manipulating of 3D-surface meshes
and semi-automated placement of surface landmarks.
}
\details{
\tabular{ll}{
Package: \tab Morpho\cr
Type: \tab Package\cr
Version: \tab 2.13\cr
Date: \tab 2025-07-23\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zarquon42b/Morpho}
  \item Report bugs at \url{https://github.com/zarquon42b/Morpho/issues}
}

}
\author{
\strong{Maintainer}: Stefan Schlager \email{zarquon42@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Gregory Jefferis [contributor]
  \item Dryden Ian [copyright holder]
}

}
