% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILDS.r
\name{plot.ILDSR2}
\alias{plot.ILDSR2}
\title{Plot graphical report for ILDSR2}
\usage{
\method{plot}{ILDSR2}(x, ...)
}
\arguments{
\item{x}{output of \code{\link{ILDSR2}}}

\item{...}{additonal parameter currently not used.}
}
\description{
Plot graphical report for ILDSR2
}
