% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider3d.r
\name{plot.slider3d}
\alias{plot.slider3d}
\title{plot the result of slider3d}
\usage{
\method{plot}{slider3d}(x, cols = 2:4, pt.size = NULL,
  point = c("sphere", "point"), specimen = 1, add = TRUE, ...)
}
\arguments{
\item{x}{result of \code{slider3d} call}

\item{cols}{vector containing colors for each coordinate type cols[1]=landmarks, cols[2]=surface landmarks, cols[3]=outlines.}

\item{pt.size}{size of plotted points/spheres. If \code{point="s"}.
\code{pt.size} defines the radius of the spheres. If \code{point="p"} it
sets the variable \code{size} used in \code{point3d}.}

\item{point}{how to render landmarks.}

\item{specimen}{integer: select the specimen to plot}

\item{add}{logical: if TRUE, a new rgl window is opened.}

\item{...}{additonal, currently unused parameters}
}
\description{
plot the result of slider3d
}
