% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls2B.r
\name{plsCoVarCommonShape}
\alias{plsCoVarCommonShape}
\title{Compute the shape changes along the common axis of deformations}
\usage{
plsCoVarCommonShape(pls, i, sdcommon = 1)
}
\arguments{
\item{pls}{object of class "pls2B"}

\item{i}{integer: dimension of latent space to show shape changes for}

\item{sdcommon}{standard deviations derived from scores scaled to a consensus scale}
}
\description{
Compute the shape changes between two blocks of 2D or 3D shape coordiantes along the common axis of deformations defined by each dimension of the latent space
}
\references{
Mitterocker and Bookstein N.N.
}
