% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupPCA.r
\name{predict.bgPCA}
\alias{predict.bgPCA}
\title{Compute between-group-PC scores from new data}
\usage{
\method{predict}{bgPCA}(object, newdata, ...)
}
\arguments{
\item{object}{object of class \code{bgPCA} returned from \code{\link{groupPCA}}}

\item{newdata}{matrix or 3D array containing data in the same format as originally used to compute groupPCA}

\item{...}{currently not used.}
}
\value{
returns the between-group-PC scores for new data
}
\description{
Compute between-group-PC scores from new data
}
