% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVA.r
\name{predict.CVA}
\alias{predict.CVA}
\title{Compute CV-scores from new data}
\usage{
\method{predict}{CVA}(object, newdata, ...)
}
\arguments{
\item{object}{object of class \code{\link{CVA}}}

\item{newdata}{matrix or 3D array containing data in the same format as originally used to compute CVA}

\item{...}{currently not used.}
}
\value{
returns the CV-scores for new data
}
\description{
Compute CV-scores from new data
}
