% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls2B.r
\name{predictPLSfromData}
\alias{predictPLSfromData}
\title{predict 2 Block-PLS from new data}
\usage{
predictPLSfromData(pls, x, y, ncomp = NULL)
}
\arguments{
\item{pls}{output of pls2B}

\item{x}{data in the same format as in original pls2B (for landmarks this can be an array or a matrix and for other data a matrix of a vector)}

\item{y}{data in the same format as in original pls2B (for landmarks this can be an array or a matrix and for other data a matrix of a vector)}

\item{ncomp}{number of (latent) components to use for prediction.}
}
\value{
returns an array/matrix/vector of predictions - depending on input for computing \code{pls}
}
\description{
predict 2 Block-PLS from new data
}
\note{
either x or y must be missing
}
\examples{
##see examples in pls2B
}
\seealso{
\code{\link{pls2B}, \link{getPLSscores},\link{predictPLSfromScores}}
}
