% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls2B.r
\name{predictPLSfromScores}
\alias{predictPLSfromScores}
\title{predict data from 2-Block PLS-scores}
\usage{
predictPLSfromScores(pls, x, y)
}
\arguments{
\item{pls}{output of pls2B}

\item{x}{scores associated with dataset x in original pls2B}

\item{y}{scores associated with dataset y in original pls2B}
}
\value{
returns an array/matrix of landmarks or original values, depending on input for computing \code{pls}
}
\description{
predict data from 2-Block PLS-scores
}
\note{
either x or y must be missing. If x-scores are provided, the yscores will be estimated and the predictions calculated.
}
\seealso{
\code{\link{pls2B}, \link{getPLSscores},\link{predictPLSfromData}}
}
