\name{read.mpp}
\alias{read.mpp}
\title{Read saved pick-points from meshlab
}
\description{
imports pick points selected with meshlab}
\usage{
read.mpp(file)
}
\arguments{
  \item{file}{file to import
}
}

\value{

\item{data }{matrix containing coordinates}
  \item{info }{additional info contained in file}

}

\author{
  Stefan Schlager

}



\seealso{
\code{\link{read.pts}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
