% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiducials.r
\name{read.slicerjson}
\alias{read.slicerjson}
\title{read Landmarks from Slicer in Json format}
\usage{
read.slicerjson(x, lps2ras = FALSE, na = NULL)
}
\arguments{
\item{x}{path to json file}

\item{lps2ras}{logical: if the coordinate system is LPS and \code{lps2ras=TRUE}, the data will be rotated into the RAS space by inverting the first two dimensions using \code{\link{LPS2RAS}}.
@param na value to be replaced by NA}
}
\value{
returns matrix or list of matrices with imported landmark coordinates
}
\description{
read Landmarks from Slicer in Json format
}
