\name{readallTPS}
\alias{readallTPS}
\title{
Import landmarks and outlines from TPS files}
\description{
Imports outlines and landmarks from files generated by tpsdig2}
\usage{
readallTPS(file)
}
\arguments{
  \item{file}{A TPS-file generated by tpsdig2
  }
}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{ID }{Specimen IDs read from TPS file
}
\item{LM }{list of landmarks contained in the TPS-file
}
\item{outlines }{a list containing sublists for each specimen with all
  the outlines read from TPS file
}
  
%% ...
}
\references{http://life.bio.sunysb.edu/ee/rohlf/software.html
%% ~put references to the literature/web site here ~
}
\author{Stefan Schlager
%%  ~~who you are~~
}
\note{currently only landmarks, ID and outlines are read from the TPS-file
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{read.lmdta}}, \code{\link{read.pts}}
}
%%\examples{

%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
