% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLandmarks.csv.r
\name{readLandmarks.csv}
\alias{readLandmarks.csv}
\title{import landmark data from csv files}
\usage{
readLandmarks.csv(file, x, y = 2:4, rownames = NULL, header = TRUE,
  dec = ".", sep = ";")
}
\arguments{
\item{file}{character: path to file containing landmark data.}

\item{x}{either a vector specifiing which rows are to be imported, or
character vector containing variable names to be sought for.}

\item{y}{a vector specifiing, which columns of the speradsheet ist to be
imported.}

\item{rownames}{integer: specifies columns, where variable names are stored.}

\item{header}{logical : if spreadsheet contains header-row.}

\item{dec}{character: defines decimal sepearator.}

\item{sep}{character: defines column seperator.}
}
\value{
\item{LM }{matrix containing imported data}
\item{NAs }{vector containing rows containing NAs}
}
\description{
import landmark data from csv files
}
\seealso{
\code{\link{read.table}}
}
\author{
Stefan Schlager
}
