% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreShapes.r
\name{restoreFromPCA}
\alias{restoreFromPCA}
\title{restore original data from PCA}
\usage{
restoreFromPCA(scores, rotation, center)
}
\arguments{
\item{scores}{matrix containing the PC-scores}

\item{rotation}{matrix containing the PCs}

\item{center}{vector containing the center}
}
\description{
restore original data from PCA by reverting rotation and centering
}
\examples{
myirispca <- prcomp(iris[,1:4])
myirisRecovered <- restoreFromPCA(myirispca$x,myirispca$rotation,myirispca$center)
all.equal(myirisRecovered,as.matrix(iris[,1:4]))
}
