% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrodeform.r
\name{retroDeform3d}
\alias{retroDeform3d}
\title{symmetrize a bilateral landmark configuration}
\usage{
retroDeform3d(mat, pairedLM, hmult = 5, alpha = 0.01)
}
\arguments{
\item{mat}{matrix with bilateral landmarks}

\item{pairedLM}{2-column integer matrix with the 1st columns containing row indices of left side landmarks and 2nd column the right hand landmarks}

\item{hmult}{factor controlling the bandwidth for calculating local weights (which will be \code{hmult} * average distance between landmarks and their closest neighbour).}

\item{alpha}{factor controlling spacing along x-axis}
}
\value{
\item{deformed}{matrix containing deformed landmarks}
\item{orig}{matrix containing original landmarks}
}
\description{
symmetrize a bilateral landmark configuration by removing bending and stretching
}
\references{
Ghosh, D.; Amenta, N. & Kazhdan, M. Closed-form Blending of Local Symmetries. Computer Graphics Forum, Wiley-Blackwell, 2010, 29, 1681-1688
}
