\name{rotaxisMat}
\alias{rotaxisMat}
\title{calculate a rotation matrix around an arbitrary axis through the origin in
3D}
\usage{
  rotaxisMat(u, theta)
}
\arguments{
  \item{u}{a vector around which to rotate}

  \item{theta}{angle in radians to rotate}
}
\value{
  returns 3x3 rotation matrix
}
\description{
  calculate a rotation matrix around an arbitrary axis in
  3D
}
\references{
  http://en.wikipedia.org/wiki/Rotation_matrix
}
\seealso{
  \code{\link{rotaxis3d}}
}
\keyword{~kwd1}
\keyword{~kwd2}

