% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossp.r, R/tanplan.r
\name{crossProduct}
\alias{crossProduct}
\alias{tangentPlane}
\title{calculate the orthogonal complement of a 3D-vector}
\usage{
crossProduct(x, y)

tangentPlane(x)
}
\arguments{
\item{x}{vector of length 3.}

\item{y}{vector of length 3.}
}
\value{
tangentPlane:

crossProduct: returns a vector of length 3.
\item{y }{vector orthogonal to x}
\item{z }{vector orthogonal to x and y}
}
\description{
calculate the orthogonal complement of a 3D-vector
}
\details{
calculate the orthogonal complement of a 3D-vector or the 3D-crossproduct,
finding an orthogonal vector to a plane in 3D.
}
\examples{

require(rgl)

x <- c(1,0,0)
y <- c(0,1,0)

#example tangentPlane
z <- tangentPlane(x)
#visualize result
\dontrun{
lines3d(rbind(0, x), col=2, lwd=2)
## show complement
lines3d(rbind(z$y, 0, z$z), col=3, lwd=2)
}
# example crossProduct
z <- crossProduct(x, y)
# show x and y
\dontrun{
lines3d(rbind(x, 0, y), col=2, lwd=2)
# show z
lines3d(rbind(0, z), col=3, lwd=2)
}
}
\author{
Stefan Schlager
}
