% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tps3d.r
\name{tpsGetCoeff}
\alias{tpsGetCoeff}
\title{calculate coefficients for a thin-plate spline deformation}
\usage{
tpsGetCoeff(refmat, tarmat, lambda = 1e-05)
}
\arguments{
\item{refmat}{reference matrix - e.g. landmark configuration on a surface}

\item{tarmat}{target matrix - e.g. landmark configuration on a target
surface}

\item{lambda}{numeric: regularisation parameter of the TPS.}
}
\value{
returns an object of class "tpsCoeff" with
\item{refmat}{reference points}
\item{tarmat}{target points}
\item{coeff}{TPS-coefficients}
\item{lambda}{lambda value as specified}
}
\description{
calculate coefficients for a thin-plate spline deformation
}
\examples{
data(nose)
## define some landmarks
refind <- c(1:3,4,19:20)
## use a subset of shortnose.lm as anchor points for a TPS-deformation
reflm <- shortnose.lm[refind,]
tarlm <- reflm
##replace the landmark at the tip of the nose with that of longnose.lm
tarlm[4,] <- longnose.lm[4,]
##  deform a set of semilandmarks by applying a TPS-deformation
##  based on 5 reference points
trafo <- tpsGetCoeff(reflm, tarlm)
## apply Deformation to a mesh
deformMesh <- applyTransform(shortnose.mesh,trafo)
retroDef <- applyTransform(deformMesh,trafo,inverse=T)
}

