% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyTransform.r
\name{applyTransform}
\alias{applyTransform}
\alias{applyTransform.default}
\alias{applyTransform.matrix}
\alias{applyTransform.mesh3d}
\title{apply affine transformation to data}
\usage{
applyTransform(x, trafo, inverse)

\method{applyTransform}{matrix}(x, trafo, inverse = FALSE)

\method{applyTransform}{mesh3d}(x, trafo, inverse = FALSE)

\method{applyTransform}{default}(x, trafo, inverse = FALSE)
}
\arguments{
\item{x}{matrix or mesh3d}

\item{trafo}{4x4 transformation matrix or an object of class "tpsCoeff"}

\item{inverse}{logical: if TRUE, the inverse of the transformation is applied (for TPS coefficients have to be recomputed)}
}
\value{
the transformed object
}
\description{
apply affine transformation to data
}
\examples{
data(boneData)
rot <- rotonto(boneLM[,,1],boneLM[,,2])
trafo <- getTrafo4x4(rot)
boneLM2trafo <- applyTransform(boneLM[,,2],trafo)
}
\seealso{
\code{\link{rotonto}, link{rotmesh.onto}, \linkt{tps3d}, \link{computeTransform}}
}

