\docType{data}
\name{boneData}
\alias{boneLM}
\alias{skull_0144_ch_fe.mesh}
\title{Landmarks and a triangular mesh}
\description{
  Landmarks on the osseous human nose and a triangular mesh representing
  this structure.
}
\usage{data(boneData)
boneLM
skull_0144_ch_fe.mesh
}
\format{
 \code{boneLM}:  A 10x3x80 array containing 80 sets of 3D-landmarks placed on the human
 osseous nose.

 
 \code{skull_0144_ch_fe.mesh}:  The mesh representing the area of the first individual of \code{boneLM}
  
}


  
\keyword{datasets}

