% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morpho-package.R
\docType{data}
\name{boneData}
\alias{boneData}
\alias{boneLM}
\alias{skull_0144_ch_fe.mesh}
\title{Landmarks and a triangular mesh}
\format{\code{boneLM}: A 10x3x80 array containing 80 sets of 3D-landmarks
placed on the human osseous nose.

\code{skull_0144_ch_fe.mesh}: The mesh representing the area of the first
individual of \code{boneLM}}
\description{
Landmarks on the osseous human nose and a triangular mesh representing this
structure.
}
\keyword{datasets}
