% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cExtract.r
\name{cExtract}
\alias{cExtract}
\title{extract information about fixed landmarks, curves and patches from and atlas
generated by "landmark"}
\usage{
cExtract(pts.file)
}
\arguments{
\item{pts.file}{either a character naming the path to a pts.file or the name
of an object imported via read.pts.}
}
\value{
returns a list containing the vectors with the indices of matrix
rows belonging to the in "landmark" defined curves, patches and fix
landmarks and a matrix containing landmark coordinates.
}
\description{
After exporting the pts file of the atlas from "landmark" and importing it
into R via "read.pts" cExtract gets information which rows of the landmark
datasets belong to curves or patches.
}
\seealso{
\code{\link{read.lmdta}} ,\code{\link{read.pts}}
}
\author{
Stefan Schlager
}
