\docType{data}
\name{colors}
\alias{bone1}
\alias{bone2}
\alias{bone3}
\alias{colors}
\alias{skin1}
\alias{skin2}
\alias{skin3}
\alias{skin4}
\title{predefined colors for bone and skin}
\format{NULL}
\description{
  predefined colors for bone and skin
}
\details{
  available colors are:

  bone1

  bone2

  bone3

  skin1

  skin2

  skin3

  skin4
}
\keyword{datasets}

