% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSize.r
\name{cSize}
\alias{cSize}
\title{calculate Centroid Size for a landmark configuration}
\usage{
cSize(x)
}
\arguments{
\item{x}{k x 3 matrix containing landmark coordinates or mesh of class "mesh3d"}
}
\value{
returns Centroid size
}
\description{
calculate Centroid Size for a landmark configuration
}
\examples{

data(boneData)
cSize(boneLM[,,1])

}
