
\name{deprecated}
\alias{mc.CVA}
\alias{mc.find.outliers}
\alias{mc.permudist}
\alias{mc.permuvar}
\alias{mc.slider3d}
\alias{mc.procSym}
\alias{c.extract}
\alias{c.size}
\alias{deform.grid}
\alias{mc.permuvec}




\title{
deprecated functions of Morpho}
\description{deprecated functions of Morpho
}

\keyword{internal}
