% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.outliers.r
\name{find.outliers}
\alias{find.outliers}
\title{Graphical interface to find outliers and/or to switch mislabeld landmarks}
\usage{
find.outliers(A, color = 4, lwd = 1, lcol = 2, mahalanobis = FALSE,
  PCuse = NULL, text = TRUE, reflection = FALSE)
}
\arguments{
\item{A}{Input k x m x n real array, where k is the number of points, m is
the number of dimensions, and n is the sample size.}

\item{color}{color of Landmarks points to be plotted}

\item{lwd}{linewidth visualizing distances of the individual landmarks from
mean.}

\item{lcol}{color of lines visualizing distances of the individual landmarks
from mean.}

\item{mahalanobis}{logical: use mahalanobis distance to find outliers.}

\item{PCuse}{integer: Restrict mahalanobis distance to the first n Principal
components.}

\item{text}{logical: if \code{TRUE}, landmark labels (rownumbers) are
displayed}

\item{reflection}{logical: specify whether reflections are allowed for superimpositioning.}
}
\value{
\item{data.cleaned }{array (in original coordinate system) containing
the changes applied and outliers eliminated}
\item{outlier }{vector with integers indicating the positions in the
original array that have been marked as outliers}
\item{dist.sort }{table showing the distance to mean for each
observation - decreasing by distance}
\item{type }{what kind of distance was used}
}
\description{
Graphical interface to find outliers and/or to switch mislabeld landmarks
}
\details{
This function performs a procrustes fit and sorts all specimen according to
their distances (either Procrustes or Mahalanobis-distance) to the sample's
consensus. It provides visual help for rearranging landmarks and/or
excluding outliers.
}
\examples{

data(boneData)
## look for outliers using the mahalanobis distance based on the first
# 10 PCscores
# to perform the example below, you need,of course, uncomment the answers
\dontrun{
outliers <- find.outliers(boneLM, mahalanobis= TRUE, PCuse=10)
# n # everything is fine 
# n # proceed to next
# s # let's switch some landmarks (3 and 4)
# 3
# 4
# n # we are done
# y # yes, because now it is an outlier
# s #enough for now
}

}
\seealso{
\code{\link{typprob}},\code{\link{typprobClass}}
}
\author{
Stefan Schlager
}
