
\name{Morpho-internal}
\alias{ang}
\alias{angle.calc}
\alias{barycenter}
\alias{calcGamma}
\alias{calcTang_U_s}
\alias{calcTang_U}
\alias{CreateL}
\alias{CreateL2D}
\alias{CVA.crova}
\alias{ProcGPA}
\alias{Semislide}
\alias{RGB2Grey}
\alias{difplot.lm2D}
\alias{difplot.lm}
\alias{groupPCAcrova}
\alias{conv2backf}
\alias{meshres}
\alias{meshres.mesh3d}
\alias{pred.PC}
\alias{pred.PCmult}
\alias{read.obj}
\alias{restrict}
\alias{orp}
\alias{cSize}
\alias{closemat}
\alias{vecx}
\alias{fx}
\alias{fq}
\alias{unbias}

\alias{rot.proc}
\alias{projBack}
\alias{setOMP}
\alias{mcNNindex}
\alias{write.obj}
%to be moved to other packages
% spinimage
\alias{spinmap}
\alias{spinimageMesh}
\alias{spinimageLM}
\alias{plotspin}
\alias{spinimage}

%mesheR
\alias{add.faces}
\alias{del.faces}
\alias{find.faces}
\alias{minClost}

%curvature
\alias{mesh2ply2}
\alias{plotCurves}
\alias{readCurves}
\alias{file2ply2}


%copies from shapes package
\alias{scaleproc}
\alias{gdif}
\alias{addo}

\title{
  low level functions of Morpho}
\description{Internal function(s)}
\keyword{internal}



