\name{Morpho-internal}
\alias{ang}
\alias{angle.calc}
\alias{barycenter}
\alias{bindArr}
\alias{calcGamma}
\alias{calcTang_U_s}
\alias{calcTang_U}
\alias{CreateL}
\alias{CreateL2D}
\alias{CVA.crova}
\alias{ProcGPA}
\alias{Semislide}
\alias{RGB2Grey}
\alias{difplot.lm2D}
\alias{difplot.lm}
\alias{groupPCAcrova}
\alias{meshres}
\alias{meshres.mesh3d}
\alias{eigenPCA}
\alias{pred.PC}
\alias{pred.PCmult}
\alias{read.obj}
\alias{restrict}
\alias{orp}
\alias{cSize}
\alias{closemat}
\alias{fx}
\alias{meshcube}
\alias{rot.proc}
\alias{projBack}
\alias{mcNNindex}
\alias{write.obj}

\alias{scaleproc}
\alias{gdif}
\alias{addo}

\title{
  low level functions of Morpho}
\description{Internal function(s)}
\keyword{internal}



