\name{Morpho-internal}
\alias{ProcGPA}

%\alias{eigenPCA}
\alias{pred.PC}
\alias{pred.PCmult}
%\alias{read.obj}
%\alias{orp}
%\alias{closemat}
%\alias{meshcube}
%\alias{rot.proc}
%\alias{projBack}
%\alias{mcNNindex}
%\alias{write.obj}

%\alias{scaleproc}
%\alias{gdif}
%\alias{addo}

\title{
  low level functions of Morpho}
\description{Internal function(s)}
\keyword{internal}



