% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv2backf.r
\name{invertFaces}
\alias{invertFaces}
\title{invert faces' orientation of triangular mesh}
\usage{
invertFaces(mesh)
}
\arguments{
\item{mesh}{triangular mesh of class \code{mesh3d}}
}
\value{
returns resulting mesh
}
\description{
inverts faces' orientation of triangular mesh and recomputes vertex normals
}
\examples{


data(nose)
\dontrun{
rgl::shade3d(shortnose.mesh,col=3)
}
noseinvert <- invertFaces(shortnose.mesh)
## show normals
\dontrun{
plotNormals(noseinvert,long=0.01)
}
}
\seealso{
\code{\link{updateNormals}}
}
\author{
Stefan Schlager
}
