% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMeshes.r
\name{mergeMeshes}
\alias{mergeMeshes}
\title{merge multiple triangular meshes into a single one}
\usage{
mergeMeshes(...)
}
\arguments{
\item{\dots}{triangular meshes of class \code{'mesh3d'} to merge or a list
of triangular meshes.}
}
\value{
returns the meshes merged into a single one.
}
\description{
merge multiple triangular meshes into a single one, preserving color and
vertex normals.
}
\examples{

require(rgl)
data(boneData)
data(nose)
mergedMesh <- mergeMeshes(shortnose.mesh, skull_0144_ch_fe.mesh)
\dontrun{
require(rgl)
shade3d(mergedMesh, col=3)
}

}
\seealso{
\code{\link{mesh2ply}, \link{file2mesh}, \link{ply2mesh}}
}
