% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh2grey.r
\name{mesh2grey}
\alias{mesh2grey}
\title{convert a colored mesh to greyscale.}
\usage{
mesh2grey(mesh)
}
\arguments{
\item{mesh}{Object of class mesh3d}
}
\value{
returns a mesh with material$color replaced by greyscale rgb values.
}
\description{
convert the colors of a colored mesh to greyscale values
}
\seealso{
\code{\link{ply2mesh}},\code{\link{file2mesh}}
}
\author{
Stefan Schlager
}
