% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deprecated}
\alias{adnormals}
\alias{conv2backf}
\alias{crossp}
\alias{deform.grid}
\alias{deprecated}
\alias{regdist.raw}
\alias{tanplan}
\title{deprecated functions of Morpho}
\usage{
deform.grid(...)

adnormals(...)

regdist.raw(dataarray, plot = TRUE, main = "", rho = "angle",
  dist.mat.out = FALSE)

crossp(...)

tanplan(...)

conv2backf(...)
}
\description{
document deprecated functions
}
\keyword{internal}

