\name{deform.grid}
\alias{deform.grid}
\alias{deprecated}
\alias{regdist.raw}
\title{deprecated functions of Morpho}
\usage{
  deform.grid(...)

  regdist.raw(dataarray, plot = TRUE, main = "",
    rho = "angle", dist.mat.out = FALSE)
}
\description{
  deprecated functions of Morpho
}
\keyword{internal}

