% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morpho-package.R
\name{deprecated}
\alias{deprecated}
\alias{deform.grid}
\alias{adnormals}
\alias{regdist.raw}
\alias{crossp}
\alias{tanplan}
\alias{conv2backf}
\alias{warp.mesh}
\title{deprecated functions of Morpho}
\usage{
deform.grid(...)

adnormals(...)

regdist.raw(dataarray, plot = TRUE, main = "", rho = "angle",
  dist.mat.out = FALSE)

crossp(...)

tanplan(...)

conv2backf(...)

warp.mesh(mesh, matr, matt, lambda = 1e-08, updateNormals = TRUE,
  silent = FALSE)
}
\description{
document deprecated functions
}
\keyword{internal}
