\name{Morpho-package}
\Rdversion{1.1}
\alias{Morpho-package}
\alias{Morpho}
\docType{package}
\title{
loads of stuff relating to Geometric Morphometrics and shape analysis
}
\description{
A toolbox for Morphometric calculations. Including sliding operations
for Semilandmarks, importing, exporting and manipulating 3D-surface
meshes. For gaining full functionality, the command line tools provided
on
\url{http://sourceforge.net/projects/morpho-rpackage/files/Auxiliaries/}
have to be installed.
}
\details{
  \tabular{ll}{
    Package: \tab Morpho\cr
    Type: \tab Package\cr
    Version: \tab 0.99\cr
    Date: \tab 2013-09-20\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr }

}
\author{Stefan Schlager


Maintainer: Stefan Schlager <stefan.schlager@uniklinik-freiburg.de>

}
\references{
Schlager S. 2013. Soft-tissue reconstruction of the human nose - population
differences and sexual dimorphism.}

\keyword{ package }


