\name{Morpho-package}
\Rdversion{1.1}
\alias{Morpho-package}
\alias{Morpho}
\docType{package}
\title{
loads of stuff relating to Geometric Morphometrics and shape analysis
}
\description{
A toolbox for Morphometric calculations. Including sliding operations
for Semilandmarks, importing, exporting and manipulating 3D-surface
meshes. For gaining full functionality, the command line tools provided
on
\url{http://sourceforge.net/projects/morpho-rpackage/files/Auxiliaries/}
have to be installed.
}
\details{
\tabular{ll}{
Package: Morpho
Type: Package
Version: 0.25-00
Date: 2013-07-19
License: GPL
LazyLoad: yes
}

}
\author{Stefan Schlager


Maintainer: Stefan Schlager <stefan.schlager@uniklinik-freiburg.de>

}
\references{
To be announced
}

\keyword{ package }


