\name{Morpho-package}
\Rdversion{1.1}
\alias{Morpho-package}
\alias{Morpho}
\docType{package}
\title{
loads of stuff relating to Geometric Morphometrics and shape analysis
}
\description{
A toolbox for Morphometric calculations. Including sliding operations
for Semilandmarks, importing, exporting and manipulating 3D-surface meshes
}
\details{
\tabular{ll}{
Package: \tab Morpho\cr
Type: \tab Package\cr
Version: \tab 0.13\cr
Date: \tab 2012-01-19\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{Stefan Schlager


Maintainer: Stefan Schlager <stefan.schlager@uniklinik-freiburg.de>

}
\references{
To be announced
}

\keyword{ package }


