\name{Morpho-package}
\Rdversion{1.1}
\alias{Morpho-package}
\alias{Morpho}
\docType{package}
\title{
loads of stuff relating to Geometric Morphometrics
}
\description{
A toolbox for Morphometric calculations. Including sliding operations for Semilandmarks and Warping of .Obj meshs.
}
\details{
\tabular{ll}{
Package: \tab Morpho\cr
Type: \tab Package\cr
Version: \tab 0.6.5.2\cr
Date: \tab 2010-07-21\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{Stefan Schlager


Maintainer: Stefan Schlager <stefan.schlager@uniklinik-freiburg.de>

}
\references{
To be announced
}

\keyword{ package }


