\docType{package}
\encoding{utf8}
\name{Morpho-package}
\alias{Morpho}
\alias{Morpho-package}
\title{A toolbox providing methods for data-acquisitiopn, visualisation and
statistical methods related to Geometric Morphometrics and shape analysis}
\description{
A toolbox for Morphometric calculations. Including sliding
operations for Semilandmarks, importing, exporting and
manipulating of 3D-surface meshes and semi-automated
placement of surface landmarks. For gaining full
functionality, the command line tools provided on
\url{http://sourceforge.net/projects/morpho-rpackage/files/Auxiliaries/}
have to be installed.
}
\details{
\tabular{ll}{ Package: \tab Morpho\cr Type: \tab Package\cr
Version: \tab 1.99.140122\cr Date: \tab 2014-01-22\cr
License: \tab GPL\cr LazyLoad: \tab yes\cr }
}
\note{
For full functionality please install trimesh-tools as
described here:
\url{http://sourceforge.net/p/morpho-rpackage/wiki/Installation_Morpho/#installation-of-the-command-line-tools-strongly-recommended}

The pdf-version of Morpho-help can be obtained from CRAN on
\url{http://cran.r-project.org/web/packages/Morpho/Morpho.pdf}
}
\author{
Stefan Schlager

Maintainer: Stefan Schlager
<stefan.schlager@uniklinik-freiburg.de>
}
\references{
Schlager S. 2013. Soft-tissue reconstruction of the human
nose: population differences and sexual dimorphism. PhD
thesis,
\enc{Universitätsbibliothek}{Universitaetsbibliothek}
Freiburg. URL:
\url{http://www.freidok.uni-freiburg.de/volltexte/9181/}.
}
\keyword{package}

