% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morpho-package.R
\docType{package}
\encoding{utf8}
\name{Morpho-package}
\alias{Morpho-package}
\alias{Morpho}
\title{A toolbox providing methods for data-acquisition, visualisation and
statistical methods related to Geometric Morphometrics and shape analysis}
\description{
A toolbox for Morphometric calculations. Including sliding operations for
Semilandmarks, importing, exporting and manipulating of 3D-surface meshes
and semi-automated placement of surface landmarks.
}
\details{
\tabular{ll}{
Package: \tab Morpho\cr
Type: \tab Package\cr
Version: \tab 2.8\cr
Date: \tab 2020-02-26\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\note{
The pdf-version of Morpho-help can be obtained from CRAN on \url{https://cran.r-project.org/package=Morpho}

For more advanced operations on triangular surface meshes, check out my package Rvcg: \url{https://cran.r-project.org/package=Rvcg} or the code repository on github \url{https://github.com/zarquon42b/Rvcg}
}
\references{
Schlager S. 2013. Soft-tissue reconstruction of the human nose:
population differences and sexual dimorphism. PhD thesis,
\enc{Universitätsbibliothek}{Universitaetsbibliothek} Freiburg. URL:
\url{http://www.freidok.uni-freiburg.de/volltexte/9181/}.
}
\author{
Stefan Schlager \email{zarquon42@gmail.com}

Maintainer: Stefan Schlager \email{zarquon42@gmail.com}
}
\keyword{package}
