% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morpho-package.R
\docType{data}
\name{nose}
\alias{nose}
\alias{shortnose.mesh}
\alias{shortnose.lm}
\alias{longnose.lm}
\title{landmarks and a triangular mesh representing a human nose}
\format{\code{shortnose.mesh}: A triangular mesh of class 'mesh3d'.

\code{shortnose.lm}: matrix containing example landmark data placed on
\code{shortnose.mesh}.

\code{longnose.lm}: matrix containing example landmark data representing a
caricaturesquely deformed human nose.}
\description{
triangular mesh representing a human nose and two matrices containing
landmark data
}
\keyword{datasets}
