% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relwarps.r
\name{predictRelWarps}
\alias{predictRelWarps}
\title{predict relative warps for data not included in the training data set}
\usage{
predictRelWarps(x, newdata)
}
\arguments{
\item{x}{output from \code{relWarps}}

\item{newdata}{k x m x n array holding new landmark data}
}
\value{
returns relative warp scores
}
\description{
predict relative warps for data not included in the training data set
}

