% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts2meshICP.R
\name{pts2meshICP}
\alias{pts2meshICP}
\title{Align a landmark configuration to a mesh using ICP}
\usage{
pts2meshICP(
  x,
  maxiter = 100,
  mindist = 1e+15,
  minchange = 1e-04,
  type = c("rigid", "similarity", "affine"),
  minscale = 0.9,
  mincumscale = NULL
)
}
\arguments{
\item{x}{moving points}

\item{maxiter}{integer: maximum number of iterations}

\item{mindist}{restrict valid points to be within this distance}

\item{minchange}{double: smallest maximum change between ICP interations}

\item{type}{character: select the transform to be applied, can be "rigid","similarity" or "affine"}

\item{minscale}{double: for types "similarity" and "affine", minimum scale factor per iteration, prevents excessive shrinking in first iteration.}

\item{mincumscale}{double: for types "similarity" and "affine", minimum cumulative scale factor, prevents shrinking below threshold (e.g. 0.7), values above 1 will enforce size increase, should only be used if both landmark configurations are in the same scale.}

\item{y}{target mesh}
}
\value{
landmark array: moving points after the alignment
}
\description{
Align a landmark configuration/pointcloud to a mesh using Iterative Closest Point search
}
\examples{
print("pseudotest")
}
