% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiducials.r
\name{read.fcsv}
\alias{read.fcsv}
\title{read fiducials from slicer4}
\usage{
read.fcsv(x, na = NULL)
}
\arguments{
\item{x}{filename}

\item{na}{value to be replaced by NA}
}
\value{
a k x 3 matrix with landmarks
}
\description{
read fiducials from slicer4
}
