% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mpp.r
\name{read.mpp}
\alias{read.mpp}
\title{Read saved pick-points from meshlab}
\usage{
read.mpp(file, info = FALSE)
}
\arguments{
\item{file}{file to import}

\item{info}{logical: if TRUE, addtional infos are returned}
}
\value{
if \code{info=FALSE}:

a matrix containing picked-points coordinates (only those tagged as active).

if \code{info=TRUE}: a list containing
\item{data }{matrix containing coordinates - including points tagged as inactive}
\item{info }{additional info contained in file.}
}
\description{
imports pick points selected with meshlab
}
\seealso{
\code{\link{read.pts}}
}
\author{
Stefan Schlager
}
