% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.pts.r
\name{read.pts}
\alias{read.pts}
\title{reads pts files}
\usage{
read.pts(file = "x", na = 9999)
}
\arguments{
\item{file}{pts file}

\item{na}{specifies a value that indicates missing values}
}
\value{
\item{matrix }{matrix containing landmark information rownames will be
the names given to the landmarks in Landmark}
}
\description{
reads Landmark data exported from the software Landmark from
http://graphics.idav.ucdavis.edu/research/EvoMorph
}
\examples{

data(nose)
write.pts(shortnose.lm, filename="shortnose")
data <- read.pts("shortnose.pts")

}
\seealso{
\code{\link{read.pts}}
}
