% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiducials.r
\name{read.slicerjson}
\alias{read.slicerjson}
\title{read Landmarks from Slicer in Json format}
\usage{
read.slicerjson(x)
}
\arguments{
\item{x}{path to json file

returns matrix or list of matrices with imported landmark coordinates}
}
\description{
read Landmarks from Slicer in Json format
}
