\name{readallTPS}
\alias{readallTPS}
\title{
Import landmarks and outlines from TPS files}
\description{
Imports outlines and landmarks from files generated by tpsdig2}
\usage{
readallTPS(file)
}
\arguments{
  \item{file}{A TPS-file generated by tpsdig2
  }
}

\value{
\item{ID }{Specimen IDs read from TPS file
}
\item{LM }{list of landmarks contained in the TPS-file
}
\item{outlines }{a list containing sublists for each specimen with all
  the outlines read from TPS file
}
  
}
\references{http://life.bio.sunysb.edu/ee/rohlf/software.html
}
\author{Stefan Schlager
}
\note{currently only landmarks, ID and outlines are read from the TPS-file
}


\seealso{
\code{\link{read.lmdta}}, \code{\link{read.pts}}
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
