% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readallTPS.r
\name{readallTPS}
\alias{readallTPS}
\title{Import landmarks and outlines from TPS files}
\usage{
readallTPS(file)
}
\arguments{
\item{file}{A TPS-file generated by tpsdig2}
}
\value{
\item{ID }{Specimen IDs read from TPS file}
\item{LM }{list of landmarks contained in the TPS-file}
\item{outlines }{a list containing sublists for each specimen with all
the outlines read from TPS file}
}
\description{
Imports outlines and landmarks from files generated by tpsdig2
}
\note{
currently only landmarks, ID and outlines are read from the TPS-file
}
\references{
http://life.bio.sunysb.edu/ee/rohlf/software.html
}
\seealso{
\code{\link{read.lmdta}}, \code{\link{read.pts}}
}
\author{
Stefan Schlager
}
