% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsOnOutline.r
\name{resampleCurve}
\alias{resampleCurve}
\title{Resample a curve equidistantly}
\usage{
resampleCurve(x, n)
}
\arguments{
\item{x}{matrix containing coordinates}

\item{n}{number of resulting points on the resampled curve}
}
\value{
returns a matrix containing the resampled curve
}
\description{
Resample a curve equidistantly (without any interpolation)
}
\examples{
data(nose)
x <- shortnose.lm[c(304:323),]
xsample <- resampleCurve(x,n=50)
}
