% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotaxis.r
\name{rotaxisMat}
\alias{rotaxisMat}
\title{calculate a rotation matrix around an arbitrary axis through the origin in
3D}
\usage{
rotaxisMat(u, theta, homogeneous = FALSE)
}
\arguments{
\item{u}{a vector around which to rotate}

\item{theta}{angle in radians to rotate}

\item{homogeneous}{logical: if TRUE a 4x4 rotation matrix is returned}
}
\value{
returns 3x3 rotation matrix
}
\description{
calculate a rotation matrix around an arbitrary axis in 3D
}
\references{
http://en.wikipedia.org/wiki/Rotation_matrix
}
\seealso{
\code{\link{rotaxis3d}}
}
