\name{needForDocu}
\alias{BEdetect}
\alias{anonymize}
\alias{NNshapeReg}
\alias{checkLM}
\alias{place.patch}
\alias{pls2B}
\alias{procAOV}
\alias{r2morphoj}
\alias{r2morphologika}
\alias{relax.mesh}
\alias{smoothMesh3d}
\alias{warpRestrict}
\alias{mcClosemat}
\alias{get.obj.size}
%dataset
\alias{longnose.lm}
\alias{shortnose.lm}
\alias{shortnose.mesh}

\title{
To be documented soon ;)}
\description{To be documented soon}

\keyword{internal}
