\name{needForDocu}
\alias{BEdetect}
\alias{adnormals}
\alias{anonymize}
\alias{NNshapeReg}
\alias{bone1}
\alias{bone2}
\alias{bone2skin}
\alias{bone3}
\alias{check.lm}
\alias{exVar}
\alias{name2factor}
\alias{name2num}
\alias{place.patch}
\alias{pls2B}
\alias{procAOV}
\alias{qqmat}
\alias{r2morphoj}
\alias{relax.mesh}
\alias{rotonmat}
\alias{showPC}
\alias{signDist}
\alias{skin1}
\alias{skin2}
\alias{skin3}
\alias{skin4}
\alias{smoothMesh3d}
\alias{warpRestrict}
\alias{lineplot}
\alias{ray2mesh}
\alias{map.mesh}
\alias{mcClosemat}
\alias{get.obj.size}
%dataset
\alias{longnose.lm}
\alias{shortnose.lm}
\alias{shortnose.mesh}

\title{
To be documented soon ;)}
\description{To be documented soon}

\keyword{internal}
