% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tps3d.r
\name{tps3d}
\alias{tps3d}
\title{thin plate spline mapping}
\usage{
tps3d(M, refmat, tarmat, lambda = 1e-05)
}
\arguments{
\item{M}{datamatrix - e.g. the matrix information of vertices of a given
surface}

\item{refmat}{reference matrix - e.g. landmark configuration on a surface}

\item{tarmat}{target matrix - e.g. landmark configuration on a target
surface}

\item{lambda}{numeric: regularisation parameter of the TPS.}
}
\value{
returns the warped datamatrix
}
\description{
maps a datamatrix via thin plate spline between calculated by a reference on
a target configuration in 2D and 3D
}
\examples{
require(Morpho)
data(nose)
## define some landmarks
refind <- c(1:3,4,19:20)
## use a subset of shortnose.lm as anchor points for a TPS-deformation
reflm <- shortnose.lm[refind,]
tarlm <- reflm
##replace the landmark at the tip of the nose with that of longnose.lm
tarlm[4,] <- longnose.lm[4,]
##  deform a set of semilandmarks by applying a TPS-deformation
##  based on 5 reference points
deformed <- tps3d(shortnose.lm, reflm, tarlm)
\dontrun{
##visualize results by applying a deformation grid
deformGrid3d(shortnose.lm,deformed,ngrid = 5)
}
}
\author{
Stefan Schlager
}
\references{
Bookstein FL. 1989. Principal Warps: Thin-plate splines and the
decomposition of deformations. IEEE Transactions on pattern analysis and
machine intelligence 11(6).
}
\seealso{
\code{\link{warp.mesh}}
}

