% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiducials.r
\name{write.fcsv}
\alias{write.fcsv}
\title{write fiducials in slicer4 format}
\usage{
write.fcsv(x, filename = dataname, description = NULL)
}
\arguments{
\item{x}{matrix with row containing 2D or 3D coordinates}

\item{filename}{will be substituted with ".fcsv"}

\item{description}{optional: character vector containing a description for each landmark}
}
\description{
write fiducials in slicer4 format
}
\examples{
require(Rvcg)
data(dummyhead)
write.fcsv(dummyhead.lm)
## remove file
unlink("dummyhead.lm.fcsv")
}
